feat(cuenta): Implementar carga de saldo en cuenta
Se añade la funcionalidad para que un usuario pueda cargar dinero en su cuenta, cumpliendo con el requisito del "Paso 1.4".

- Se crea el endpoint `PUT /cuentas/{idCuenta}/cargar-saldo`.
- El endpoint recibe un cuerpo JSON con el monto a cargar (ej. `{"monto": 500.0}`).
- Se implementa el método `cargarSaldo` en `CuentaService` que busca la cuenta, valida que el monto sea positivo, suma el valor al saldo actual y guarda el resultado.
- Se añade manejo de errores para montos inválidos o si la cuenta no se encuentra.

+++ Integrador4/cuenta/src/main/java/grupo4/mscvcuenta/entity/Cuenta.java
@@ -23,7 +23,7 @@ public class Cuenta {
     private String usuarios;
 
     @Column(name = "monto", nullable = false)
-    private BigDecimal monto;
+    private Double monto;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "estado", nullable = false)
